Upgraded the used libraries and dependencies:
- removed the old compendium bundle (4.2) in favor of 4.3;
- fixed a couple of compiler warnings that were caused by this upgrade;
- when running the itests on the latest Jetty seems to cause subtle timing
  issues in the setup of the integration tests causing them to fail. A
  workaround seems to add a small delay between the setup of the provisioned
  services and the additional services.



git-svn-id: https://svn.apache.org/repos/asf/ace/trunk@1727044 13f79535-47bb-0310-9956-ffa450edef68

+++ org.apache.ace.authentication.itest/src/org/apache/ace/it/authentication/ObrAuthenticationTest.java
@@ -148,11 +148,13 @@ public class ObrAuthenticationTest extends AuthenticationTestBase {
 
     @Override
     public void doTearDown() throws Exception {
-        FileUtils.removeDirectoryWithContent(m_storeLocation);
-        Configuration configuration = getConfiguration(m_authConfigPID);
-        if (configuration != null) {
-            configuration.delete();
+        if (m_authConfigPID != null) {
+            Configuration configuration = getConfiguration(m_authConfigPID);
+            if (configuration != null) {
+                configuration.delete();
+            }
         }
+        FileUtils.removeDirectoryWithContent(m_storeLocation);
     }
 
     /**
